♻️ Front / Seperate board compoent and implement board-list component
