+++ front/src/components/board/board.scss
@@ -7,75 +7,75 @@
     margin-bottom: 24px;
   }
 
-  &-list {
-    @include column-flexbox(center, start);
-    width:100%;
-  }
+  // &-list {
+  //   @include column-flexbox(center, start);
+  //   width:100%;
+  // }
 
-  &-pagination {
-    @include flexbox;
-    background-color: $background;
-    position: relative;
-    padding: 20px 10px;
+  // &-pagination {
+  //   @include flexbox;
+  //   background-color: $background;
+  //   position: relative;
+  //   padding: 20px 10px;
 
-    .page-button {
-      background-color: $white;
-      width: 24px;
-      height: 24px;
-      margin-right: 6px;
-      font-size: 16px;
-      font-weight: 600;
-      border: 1px solid $border-hover;
-      border-radius: 50%;
-      transition: transform 100ms ease-in-out;
+  //   .page-button {
+  //     background-color: $white;
+  //     width: 24px;
+  //     height: 24px;
+  //     margin-right: 6px;
+  //     font-size: 16px;
+  //     font-weight: 600;
+  //     border: 1px solid $border-hover;
+  //     border-radius: 50%;
+  //     transition: transform 100ms ease-in-out;
 
-      &:last-child {
-        margin-right: 0;
-      }
+  //     &:last-child {
+  //       margin-right: 0;
+  //     }
 
-      &:not(.icon-button):hover {
-        transform: scale(1.1);
-      }
-    }
+  //     &:not(.icon-button):hover {
+  //       transform: scale(1.1);
+  //     }
+  //   }
 
-    .board-write-button {
-      position:absolute;
-      top: 50%;
-      right: 5px;
-      transform: translateY(-50%);
+  //   .board-write-button {
+  //     position:absolute;
+  //     top: 50%;
+  //     right: 5px;
+  //     transform: translateY(-50%);
 
-      font-size: 14px;
-      font-weight: 600;
-      background-color: $border-hover;
-      padding: 5px 10px;
-      border-radius: 4px;
-      color: $white;
-      transition: opacity 100ms ease-in-out;
+  //     font-size: 14px;
+  //     font-weight: 600;
+  //     background-color: $border-hover;
+  //     padding: 5px 10px;
+  //     border-radius: 4px;
+  //     color: $white;
+  //     transition: opacity 100ms ease-in-out;
 
-      &:hover {
-        opacity: 0.8;
-      }
-    }
-  }
+  //     &:hover {
+  //       opacity: 0.8;
+  //     }
+  //   }
+  // }
 
   @include responsive(D) {
     padding-top: 30px;
     
-    &-pagination {
-      padding: 30px 10px;
+  //   &-pagination {
+  //     padding: 30px 10px;
 
-      .page-button {
-        width: 30px;
-        height: 30px;
-        margin-right: 10px;
-        font-size: 18px;
-      }
+  //     .page-button {
+  //       width: 30px;
+  //       height: 30px;
+  //       margin-right: 10px;
+  //       font-size: 18px;
+  //     }
   
-      .board-write-button {
-        right: 10px;
-        font-size: 20px;
-        padding: 8px 12px;
-      }
-    }
+  //     .board-write-button {
+  //       right: 10px;
+  //       font-size: 20px;
+  //       padding: 8px 12px;
+  //     }
+  //   }
   }
 }
