+++ front/src/styles/main.scss
@@ -18,4 +18,5 @@
 @import '../components/product-item/product-item.scss';
 @import '../components/board-item/board-item.scss';
 @import '../components/board/board.scss';
+@import '../components/board-list/board-list.scss';
 @import '../components/board-write/board-write.scss';
